last modification

+++ back-end/controllers/books.js
@@ -146,23 +146,27 @@ exports.modifyBook = (req, res, next) => {
 };
 
 
-exports.deleteBook = (req, res, next) => {
-  Book.findOne({ _id: req.params.id })
-    .then(book => {
-      if (book.userId != req.auth.userId) {
-        res.status(403).json({ message: '403: unauthorized request' });
-      } else {
-        const filename = book.imageUrl.split('/pictures/')[1];
-        fs.unlink(`pictures/${filename}`, () => {
-          Book.deleteOne({ _id: req.params.id })
-            .then(() => { res.status(200).json({ message: 'Livre supprimé !' }); })
-            .catch(error => res.status(400).json({ error }));
-        });
-      }
-    })
-    .catch(error => {
-      res.status(404).json({ error });
-    });
+exports.deleteBook = async (req, res, next) => {
+  try {
+    const book = await Book.findOne({ _id: req.params.id });
+
+    if (!book) {
+      return res.status(404).json({ message: 'Livre non trouvé' });
+    }
+
+    if (book.userId != req.auth.userId) {
+      return res.status(403).json({ message: '403: unauthorized request' });
+    }
+
+    const filename = book.imageUrl.split('/pictures/')[1];
+    await fs.unlink(`pictures/${filename}`); // Suppression de l'image avec fs.promises.unlink
+
+    await Book.deleteOne({ _id: req.params.id });
+
+    return res.status(200).json({ message: 'Livre supprimé !' });
+  } catch (error) {
+    return res.status(500).json({ error: 'Erreur lors de la suppression du livre' });
+  }
 };
 
 exports.addRating = async (req, res) => {
